+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the Viewer Metrics Chrome Extension will be documented in this file.
 
+## [0.9.94] - 2025-12-09
+- Viewerlist page now shows top botted months and days
+- Percentage of accounts created after 2020 shown
+- Fixed date formatting
+- Minor fixes
+
 ## [0.9.93] - 2025-12-09
 - Export of viewer tracking data (CSV, XML, SQL, JSON)
 - Export of viewer graph history data (CSV, XML, SQL, JSON)